# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MBN AI 트롯 팬덤 플랫폼 API
  version: 0.1.0
  description: |
    FE/BE 공통 계약. **이 파일이 단일 진실 공급원입니다.**

    - BE: 이 스펙대로 컨트롤러를 구현합니다. 스펙에 없는 엔드포인트를 임의로 추가하지 마세요.
    - FE: `npm run api:types` 로 이 파일에서 TS 타입을 생성합니다. 타입을 손으로 쓰지 마세요.
    - 계약 변경은 **이 파일 수정 → 양쪽 반영** 순서로만 진행합니다.

    도메인 정의는 `docs/domain-model.md` 참조.
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: /
    description: 동일 오리진 배포
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/contents:
    get:
      tags: [Content]
      summary: 콘텐츠 목록 (기사·영상, 최신순)
      security: []
      parameters:
        - $ref: '#/components/parameters/StarIdQuery'
        - name: type
          in: query
          description: 미지정 시 기사·영상을 섞어서 반환합니다 (HOME 아카이브 캐러셀).
          schema: { $ref: '#/components/schemas/ContentType' }
        - name: live
          in: query
          description: true면 LIVE 배지가 붙은 콘텐츠만.
          schema: { type: boolean }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/ContentSummary' }
  /api/v1/contents/{id}:
    get:
      tags: [Content]
      summary: 콘텐츠 상세 (기사 또는 영상)
      description: |
        `type` 값에 따라 채워지는 필드가 다릅니다.
        - `ARTICLE` — `body`, `reporterName`, `reporterAvatarUrl`, `places`
        - `VIDEO` — `mediaUrl`, `durationSec`, `live`, `viewerCount`

        로그인 상태면 `liked` / `subscribed` 가 실제 값으로 내려옵니다.
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Content' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/contents/{id}/related:
    get:
      tags: [Content]
      summary: 관련 콘텐츠
      description: 상세 화면 하단 `관련 콘텐츠` / `관련 영상` 캐러셀.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: limit
          in: query
          schema: { type: integer, default: 6, maximum: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ContentSummary' }
...
  /api/v1/contents/{id}/ai-analysis:
    get:
      tags: [AiAnalysis]
      summary: 콘텐츠 AI 분석 조회
      description: |
        기사·영상 상세의 AI 분석 패널.
        **사전 생성된 결과를 DB에서 읽습니다** — 요청 시점에 LLM을 호출하지 않습니다.
        해당 언어의 분석이 없으면 404 대신 `ko` 결과를 폴백으로 반환합니다.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiAnalysis' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/ai-analysis/regenerate:
    post:
      tags: [AiAnalysis]
      summary: AI 분석 재생성 (관리자)
      description: 시연 중 라이브 생성을 보여주기 위한 엔드포인트입니다.
      parameters:
        - name: contentId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: locale
          in: query
          schema: { $ref: '#/components/schemas/Locale' }
      responses:
        '200':
          description: 재생성 완료
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiAnalysis' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503':
          description: AI 제공자 호출 실패
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ───────────────────────── Comment ──────────────────────────
  /api/v1/contents/{id}/comments:
    get:
      tags: [Comment]
      summary: 댓글 목록
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [Comment]
      summary: 댓글 작성
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, minLength: 1, maxLength: 500 }
      responses:
        '201':
          description: 작성됨
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/comments/{id}:
    delete:
      tags: [Comment]
      summary: 댓글 삭제 (본인만)
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '204': { description: 삭제됨 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 본인 댓글이 아님
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/comments/{id}/like:
    post:
      tags: [Reaction]
      summary: 댓글 좋아요
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LikeState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Reaction]
      summary: 댓글 좋아요 취소
      parameters: [{ $ref: '#/components/parameters/IdPath' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LikeState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/comments/{id}/translation:
    get:
      tags: [Comment]
      summary: 댓글 번역
      description: |
        댓글의 번역 아이콘에 대응합니다. 결과는 `(commentId, locale)` 단위로 캐싱되며
        두 번째 호출부터는 LLM을 거치지 않습니다.
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [commentId, locale, translatedBody]
                properties:
                  commentId: { type: integer, format: int64 }
                  locale: { $ref: '#/components/schemas/Locale' }
                  translatedBody: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '503':
          description: 번역 제공자 호출 실패
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─────────────────────── Subscription ───────────────────────
...
  /api/v1/chat/sessions:
    post:
      tags: [Chat]
      summary: AI 도우미 "비엔이" 대화 세션 생성
      description: |
        비회원도 생성 가능합니다. 세션은 스타 + 언어에 종속됩니다.
        마이크 FAB 또는 배너 CTA로 음성 오버레이를 열 때 호출합니다.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [starId]
              properties:
                starId: { type: integer, format: int64 }
                locale: { $ref: '#/components/schemas/Locale' }
      responses:
        '201':
          description: 세션 생성됨
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, suggestedQuestions]
                properties:
                  sessionId: { type: string, format: uuid }
                  suggestedQuestions:
                    type: array
                    description: 배너 진입 시 노출할 추천 질문
                    items: { type: string }
                    example:
                      - 이번 주 임영웅 출연 방송 알려줘
                      - 최근 무대 영상 추천해줘
                      - 서울에서 진행되는 팬 모임이 있어?
  /api/v1/chat/sessions/{sessionId}/messages:
    post:
      tags: [Chat]
      summary: 메시지 전송 (SSE 스트리밍 응답)
      description: |
        `Accept: text/event-stream` 으로 요청하면 토큰 단위로 스트리밍합니다.

        **SSE 이벤트 종류**
        | event | data |
        |---|---|
        | `delta` | `{"text": "부분 문자열"}` |
        | `citations` | `{"citations": [Citation]}` — 답변 근거, 딥링크용 |
        | `done` | `{"messageId": 123}` |
        | `error` | `{"code": "...", "message": "..."}` |

        `Accept: application/json` 이면 완성된 `ChatMessage` 하나를 반환합니다.

        **음성 입력은 별도 엔드포인트가 없습니다.** STT는 브라우저(Web Speech API)에서
        처리하고, 확정된 텍스트를 이 엔드포인트로 보냅니다 (`docs/ai-stack.md` §2).

        **스코프 제한**: MBN 방송·트롯 아티스트·플랫폼 내부 데이터 외의 질문에는 답하지 않고
        안내 문구를 반환합니다. 이때도 200이며 `outOfScope: true` 가 함께 옵니다.
      security:
        - {}
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        '200':
          description: OK
          content:
            text/event-stream:
              schema: { type: string }
            application/json:
              schema: { $ref: '#/components/schemas/ChatMessage' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429':
          description: 요청 한도 초과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ─────────────────────────── Auth ───────────────────────────
  # 간이 인증입니다. 회원가입·비밀번호·소셜 로그인을 구현하지 않습니다.
  # 데모 계정을 목록에서 골라 원클릭 로그인하며, 토큰은 만료가 없습니다.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
    StarIdQuery:
      name: starId
      in: query
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    AcceptLanguage:
      name: Accept-Language
      in: header
      schema: { type: string, example: ko-KR }
  responses:
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: GATHERING_ALREADY_APPLIED }
        message: { type: string, example: 이미 신청한 모임입니다. }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, last]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        last: { type: boolean }
    Locale:
      type: string
      description: |
        지원 언어 7종. 언어 선택 Bottom Sheet의 항목 순서와 대응합니다.
        **번역 검수는 KO/EN만** 합니다 — 나머지는 AI 생성 결과를 그대로 씁니다.
      enum: [EN, KO, FR, JA, ES, ZH, RU]
      default: KO
    Country:
      type: string
      description: 댓글 국가 배지에 사용합니다.
      enum: [KR, US, JP, FR, ES, CN, RU]
    LikeState:
      type: object
      required: [liked, likeCount]
      properties:
        liked: { type: boolean }
        likeCount: { type: integer }
...
    User:
      type: object
      required: [id, nickname, profileImageUrl, country, role, locale]
      properties:
        id: { type: integer, format: int64 }
        nickname: { type: string, example: 트롯사랑 }
        profileImageUrl: { type: string, format: uri, description: 댓글 아바타 }
        country: { $ref: '#/components/schemas/Country' }
        role: { type: string, enum: [MEMBER, ADMIN] }
        locale: { $ref: '#/components/schemas/Locale' }
...
    Channel:
      type: object
      required: [id, name, subscriberCount]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: MBN NEWS }
        logoUrl: { type: string, format: uri }
        subscriberCount: { type: integer }

    # ── Schedule ──
...
    ContentType:
      type: string
      enum: [ARTICLE, VIDEO]
    ContentSummary:
      type: object
      description: |
        캐러셀 카드용. `type`에 따라 카드 하단 메타 표시가 달라집니다.
        - `ARTICLE` → 좌 `channelName`, 우 `publishedAt` 상대 시각
        - `VIDEO` → 좌 `durationSec`, 우 `viewCount`
        - `VIDEO` + `live=true` → 좌 LIVE 배지, 우 `viewerCount`
      required: [id, type, title, thumbnailUrl, publishedAt, channelName, viewCount, live]
      properties:
        id: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/ContentType' }
        title: { type: string, example: 0721 무대 콘서트 직관 영상 }
        thumbnailUrl: { type: string, format: uri }
        publishedAt: { type: string, format: date-time }
        channelName: { type: string, example: MBN NEWS }
        viewCount: { type: integer }
        likeCount: { type: integer }
        commentCount: { type: integer }
        live:
          type: boolean
          description: |
            LIVE 배지 표시 여부. **실시간 스트리밍 연동이 아닙니다** — 플래그가 켜져 있으면
            배지를 붙이고 `viewerCount`를 그대로 표시할 뿐입니다.
        viewerCount:
          oneOf:
            - { type: integer }
            - { type: 'null' }
          description: live=true일 때만 값이 있습니다.
        durationSec:
          oneOf:
            - { type: integer }
            - { type: 'null' }
          description: VIDEO일 때만 값이 있습니다.
    Content:
      allOf:
        - $ref: '#/components/schemas/ContentSummary'
        - type: object
          required: [channel, liked, subscribed]
          properties:
            channel: { $ref: '#/components/schemas/Channel' }
            liked:
              type: boolean
              description: 비로그인 시 항상 false.
            subscribed:
              type: boolean
              description: 채널 구독 여부. 비로그인 시 항상 false.
            # ARTICLE 전용
            body:
              oneOf:
                - { type: string }
                - { type: 'null' }
              description: |
                기사 본문. 용어 하이라이트는 `[[용어|설명]]` 마크업으로 표기하며
                FE에서 파싱해 툴팁으로 렌더합니다. 별도 API가 없습니다.
            reporterName:
              oneOf: [{ type: string }, { type: 'null' }]
            reporterAvatarUrl:
              oneOf: [{ type: string, format: uri }, { type: 'null' }]
            places:
              type: array
              description: 뉴스 상세 하단 `기사에 나온 그 곳`. ARTICLE에서만 채워집니다.
              items: { $ref: '#/components/schemas/Place' }
            # VIDEO 전용
            mediaUrl:
              oneOf: [{ type: string, format: uri }, { type: 'null' }]

    # ── AI 분석 ──
    AiAnalysis:
      type: object
      description: 기사·영상 상세의 AI 분석 패널. 사전 생성된 결과입니다.
      required: [contentId, locale, summary, items, generatedAt]
      properties:
        contentId: { type: integer, format: int64 }
        locale: { $ref: '#/components/schemas/Locale' }
        summary: { type: string, example: AI가 기사를 분석한 내용입니다. }
        items:
          type: array
          items:
            type: object
            required: [title, body]
            properties:
              title: { type: string, example: 주요 내용 }
              body: { type: string }
        generatedAt: { type: string, format: date-time }

    # ── 댓글 ──
    Comment:
      type: object
      required: [id, contentId, author, body, likeCount, liked, createdAt]
      properties:
        id: { type: integer, format: int64 }
        contentId: { type: integer, format: int64 }
        author: { $ref: '#/components/schemas/User' }
        body: { type: string }
        likeCount: { type: integer }
        liked:
          type: boolean
          description: 비로그인 시 항상 false.
        createdAt: { type: string, format: date-time }

    # ── Gathering ──
...
    PlaceType:
      type: string
      enum: [RESTAURANT, CAFE, VENUE, FILMING_LOCATION]
    Place:
      type: object
      required: [id, name, type, address, imageUrl, sourceUrl]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, example: 한정식 서울역점 }
        type: { $ref: '#/components/schemas/PlaceType' }
        address: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        imageUrl: { type: string, format: uri }
        visitContext: { type: string, description: 스타가 방문한 프로그램 또는 시점 }
        sourceUrl:
          type: string
          format: uri
          description: '**필수** — 공개된 출처만 등록. 실시간 위치·사적 동선은 금지.'
        mapUrl:
          oneOf: [{ type: string, format: uri }, { type: 'null' }]
          description: 카카오맵 외부 링크. 지도를 임베드하지 않습니다.
        relatedContentIds:
          type: array
          items: { type: integer, format: int64 }

    # ── Tip ──
...
    Citation:
      type: object
      required: [type, id, title]
      description: AI 답변의 근거 리소스. FE는 이를 카드로 렌더링하고 해당 화면으로 딥링크합니다.
      properties:
        type: { type: string, enum: [SCHEDULE, CONTENT, GATHERING, PLACE, TIP] }
        id: { type: integer, format: int64 }
        title: { type: string }
    ChatMessage:
      type: object
      required: [id, role, content, outOfScope, createdAt]
      properties:
        id: { type: integer, format: int64 }
        role: { type: string, enum: [USER, ASSISTANT] }
        content: { type: string }
        outOfScope:
          type: boolean
          description: |
            MBN 방송·트롯 아티스트·플랫폼 데이터 밖의 질문이라 답변을 거절한 경우 true.
            FE는 이때 근거 카드를 렌더링하지 않습니다.
        citations:
          type: array
          items: { $ref: '#/components/schemas/Citation' }
        createdAt: { type: string, format: date-time }